requirements.txt
numpy==1.23.4
pandas==1.5.1
pytesseract==0.3.10
frontend==0.0.3
flask==2.2.2
matplotlib==3.6.2
pdf2image==1.16.0
pyserial==3.5
opencv-python~=4.6.0.66
pathlib~=1.0.1
requests~=2.28.1
Pillow~=9.3.0